requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.13
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mslex==1.3.0
mssql-django==1.5
psutil==6.1.1
PyJWT==2.9.0
pyodbc==5.2.0
python-decouple==3.8
pytz==2025.1
sqlparse==0.5.3
taskipy==1.14.1
tomli==2.2.1
tzdata==2025.1
